Minor fixes ported to c++
Fixes a some problems with the Frame::parse method. Namely it has been renamed to create_from_file and returns a heap object, this changed allows us to use the polymorphic behaviour not achieve with the previous method because of the static_cast.

All C code was ported to C++, namely the reading of the file.