Minor fixes and new comments
With this commit new methods were implemented in the block class, namely the operator overload of operator== and operator=(const char*). The former allows the comparison of two blocks while the later allows the copying a chunk of bytes into the block, note that the method does not check for bounds, instead it trust the caller makes sure the array is at least the size of the block.
In the video class were some fixes to the way we read the frames although is still not fully working for all types. Also we now expose a new method display that shows the whole video in a window. Another method added with this commit is one to allow to reset the position of the stream so that we can begin processing the video from the start.

The frame442 convert method also received some love but its not tested.