Many structural changes.
First of all the predictors have change quite a lot, the predict and guess methods
are now implemented in the base class predictor and the LinearPredictor and NonLinearPredictor
class are only to initialize the Predictor correctly.

The golomb class accepts a predictor from where it will encode/decode to/from the output/input file.

Addition of a cav-header header file, where the headers of the conding mechanisim will be defined,
the base CAVHeader has a paylod of 36 bytes that allows it to be generic enough (for simplicity).

The Bitstream class has been modified to write/read a CAVHeader.
Initial testing in encondeDecode utility